Moved CO2_EMISSIONS_GRAMS_PER_GB into data folder, renamed as COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB
