+++ src/__tests__/usePopupSelectedCountriesPercentage.test.ts
@@ -1,6 +1,6 @@
 import { renderHook, act } from "@testing-library/react";
 
-import { CO2_EMISSIONS_GRAMS_PER_GB } from "../constants/Countries";
+import { COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB } from "../data/constants/CountryEmissions";
 import { usePopup } from "../usePopup";
 import { mockChrome } from "./mockChrome";
 
@@ -19,10 +19,10 @@ describe("usePopup", () => {
         const mockCountry1Percentage = 0.4;
         const mockCountry2Percentage = 0.4;
         const mockAverageSpecificEmissions =
-            mockCountry1Percentage * CO2_EMISSIONS_GRAMS_PER_GB[mockCountry1] +
-            mockCountry2Percentage * CO2_EMISSIONS_GRAMS_PER_GB[mockCountry2] +
+            mockCountry1Percentage * COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB[mockCountry1] +
+            mockCountry2Percentage * COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB[mockCountry2] +
             (1 - mockCountry1Percentage - mockCountry2Percentage) *
-                CO2_EMISSIONS_GRAMS_PER_GB["World Average"];
+                COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB["World Average"];
 
         await act(async () => {
             await result.current.addSelectedCountry(mockCountry1);
