+++ src/components/CountryDropdown.tsx
@@ -2,9 +2,9 @@ import React, { Fragment } from "react";
 import { Listbox, Transition } from "@headlessui/react";
 import { PlusIcon } from "@heroicons/react/20/solid";
 import {
-    CO2_EMISSIONS_GRAMS_PER_GB,
+    COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB,
     CountryName,
-} from "../constants/Countries";
+} from "../data/constants/CountryEmissions";
 
 type CountryDropdownType = {
     addSelectedCountry: (country: CountryName) => void;
@@ -33,7 +33,7 @@ export const CountryDropdown = ({
                         leaveTo="opacity-0"
                     >
                         <Listbox.Options className="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                            {Object.entries(CO2_EMISSIONS_GRAMS_PER_GB).map(
+                            {Object.entries(COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB).map(
                                 ([country]) => (
                                     <Listbox.Option
                                         key={country}
