+++ src/usePopup.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { CountryName } from "./constants/Countries";
+import { CountryName } from "./data/constants/CountryEmissions";
 import { calculateAverageSpecificEmissionsHelper } from "./helpers/calculateAverageSpecificEmissions";
 import { calculateCarbon } from "./helpers/calculateCarbon";
 import { ISelectedCountriesRepository } from "./data/selected_countries/ISelectedCountriesRepository";
